Semantickernel llamasharp (#15)
* Restructured all projects
Moved semantic kernel and semntic memory together with new llamasharp-semantic kernel projects
Tested new mistral dolphin/and orcas models from hugginface/thebloke

* some minor Modification

* Updated all non preview nuget packages

* package update-step 2

* Updated llamasharp-semantickernel

* Upgraded packages
added Seq logginh
Updated LlamaSharp usages to conform with v0.6

* appsettings changes to logging

* Minor adjustments

+++ src/Semantic.Kernel.LlamaSharp/Tests/Semantic.Kernel.LlamaSharp.Tests/Tests/TestOfLlamaSharpSemanticKernelPrompt.cs
@@ -79,7 +79,7 @@ public class TestOfLlamaSharpSemanticKernelPrompt : IAsyncLifetime
         // Load weights into memory
         var parameters = new ModelParams(modelPath)
         {
-            Seed = RandomNumberGenerator.GetInt32(int.MaxValue),
+            Seed = (uint)RandomNumberGenerator.GetInt32(int.MaxValue),
         };
         using var model = LLamaWeights.LoadFromFile(parameters);
         var ex = new StatelessExecutor(model, parameters);
